CLN: added dummy_methods and dummy_classes libraries for force_wfmanager/tests routines
