+++ force_wfmanager/ui/setup/data_source_model_view.py
@@ -9,9 +9,9 @@ from traitsui.table_column import ObjectColumn
 from force_bdss.api import (BaseDataSourceModel, BaseDataSource, Identifier,
                             InputSlotInfo, OutputSlotInfo)
 
-from force_wfmanager.ui.view_utils import (
+from force_wfmanager.ui.ui_utils import (
     get_factory_name, get_default_background_color)
-from force_wfmanager.ui.variable_names_registry import (
+from force_wfmanager.utils.variable_names_registry import (
     VariableNamesRegistry)
