+++ force_wfmanager/utils/variable_names_registry.py
